Added default profile picture in case user does not provide one on sign up
