+++ app/controllers/api.js
@@ -89,18 +89,18 @@ export const createUser = ({
         SIGNUP_ERRORS_GENERIC,
         SIGNUP_ERRORS_MISSING_VALUES,
         SIGNUP_ERRORS_INVALID_VALUES,
-        UPLOADS_RELATIVE_PATH
+        UPLOADS_RELATIVE_PATH,
+        DEFAULT_PROFILE_PICTURE_RELATIVE_PATH,
+        JWT_SECRET
     } = process.env;
 
-    let imageFields = {};
+    // Start with the default profile picture
+    let profilePictureLink = DEFAULT_PROFILE_PICTURE_RELATIVE_PATH;
 
     if (filename && mimetype && path) {
         // We have an image upload that we need to include in the saved user
         // NOTE: Validation middleware has already run by the time we get here so we can assume the image is valid
-
-        imageFields = {
-            profilePictureLink: `${UPLOADS_RELATIVE_PATH}${filename}`
-        };
+        profilePictureLink = `${UPLOADS_RELATIVE_PATH}${filename}`
     }
 
     if (!name || !email || !password || !userType) {
@@ -164,7 +164,7 @@ export const createUser = ({
                 email,
                 password: hashedPassword,
                 isLandlord: userType === process.env.USER_TYPE_LANDLORD,
-                ...imageFields
+                profilePictureLink
             },
             returnInsertedDocument: true
         });
@@ -180,7 +180,7 @@ export const createUser = ({
     }
 
     // Now that the user has been saved, return a jwt encapsulating the new user (transformered for output)
-    const token = jwt.sign(transformUserForOutput(savedUser), process.env.JWT_SECRET);
+    const token = jwt.sign(transformUserForOutput(savedUser), JWT_SECRET);
 
     return res.json({
         token
